# refactor: ppx deriving visitors for generic AST