--- a/libs/ast_generic/dune
+++ b/libs/ast_generic/dune
@@ -14,6 +14,7 @@
       ppx_deriving.eq
       ppx_hash
       profiling.ppx
+      visitors.ppx
    )
  )
 )