--- a/semgrep.opam
+++ b/semgrep.opam
@@ -54,6 +54,7 @@ depends: [
   "cohttp-lwt-unix" # this brings lots of dependencies. This is for osemgrep.
   "lsp" {= "1.7.0"}
   "comby-kernel" {= "1.4.1"}
+  "visitors" {= "20210608"}
 ]
 
 build: [make]